Deduplicate component codes
